pipeline/src/requirements.txt
black==21.5b1
flake8>=3.7.9, <4.0
ipython~=7.0
isort~=5.0
jupyter~=1.0
jupyter_client>=5.1, <7.0
jupyterlab~=3.0
kedro[pandas.CSVDataSet]==0.17.5
kedro-telemetry~=0.1.0
nbstripout~=0.4
pytest-cov~=2.5
pytest-mock>=1.7.1, <2.0
pytest~=6.2
wheel>=0.35, <0.37
pandas>=1.0.1
numpy>=1.18.1
matplotlib>=3.1.3
tqdm>=4.42.1
networkx>=2.0
pyvis>=0.1.9
scipy>=1.4.1
mca>=1.0.3
sklearn
h2o>=3.34.0.1
psutil>=5.6.7
openpyxl>=3.0.3
supervenn
loging
reactome2py>=3.0.0
pathlib==1.0.1
Pillow
seaborn
xlrd==1.2.0
xlsxwriter
matplotlib_venn